added all required business logic and relations between entities; added a compound controller which merges all other controllers

+++ src/main/java/org/example/tpo04blog/repositories/ArticleRepository.java
@@ -3,5 +3,5 @@ package org.example.tpo04blog.repositories;
 import org.example.tpo04blog.entities.Article;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArticleRepository extends CrudRepository<Article, Integer> {
+public interface ArticleRepository extends CrudRepository<Article, Long> {
 }
