+++ src/main/java/org/example/tpo04blog/repositories/RoleRepository.java
@@ -3,5 +3,5 @@ package org.example.tpo04blog.repositories;
 import org.example.tpo04blog.entities.Role;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoleRepository extends CrudRepository<Role, Integer> {
+public interface RoleRepository extends CrudRepository<Role, Long> {
 }
